+++ routes/core/streams/index.js
@@ -8,6 +8,7 @@ const hash = require('../../../utils/misc/hash.js').hash
 const Converter = require('../../../utils/converter/converter')
 const { hasStreamPermission } = require('../../../middleware/authorization/roles')
 const { getPermissions, hasPermission, STREAM, PROJECT, READ, UPDATE } = require('../../../services/roles')
+const arbimonService = require('../../../services/arbimon')
 
 /**
  * @swagger
@@ -43,46 +44,56 @@ const { getPermissions, hasPermission, STREAM, PROJECT, READ, UPDATE } = require
  *         description: Invalid query parameters
  */
 
-router.post('/', function (req, res) {
-  const userId = req.rfcx.auth_token_info.owner_id
-  const convertedParams = {}
-  const params = new Converter(req.body, convertedParams)
-  params.convert('id').optional().toString()
-  params.convert('name').toString()
-  params.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
-  params.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
-  params.convert('altitude').optional().toFloat()
-  params.convert('description').optional().toString()
-  params.convert('is_public').optional().toBoolean().default(false)
-  params.convert('external_id').optional().toInt()
-  params.convert('project_id').optional().toString()
-  params.convert('project_external_id').optional().toInt()
+router.post('/', async (req, res) => {
+  let stream
+  try {
+    const userId = req.rfcx.auth_token_info.owner_id
+    const converter = new Converter(req.body, {})
+    const source = req.headers.source
+    converter.convert('id').optional().toString().toLowerCase()
+    converter.convert('name').toString()
+    converter.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
+    converter.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
+    converter.convert('altitude').optional().toFloat()
+    converter.convert('description').optional().toString()
+    converter.convert('is_public').optional().toBoolean().default(false)
+    converter.convert('external_id').optional().toInt()
+    converter.convert('project_id').optional().toString()
+    converter.convert('project_external_id').optional().toInt()
 
-  return params.validate()
-    .then(() => usersFusedService.ensureUserSyncedFromToken(req))
-    .then(async () => {
-      if (convertedParams.project_id) {
-        await projectsService.getById(convertedParams.project_id)
-        const allowed = await hasPermission(UPDATE, req.rfcx.auth_token_info, convertedParams.project_id, PROJECT)
-        if (!allowed) {
-          throw new ForbiddenError('You do not have permission to create stream in this project.')
-        }
+    const params = await converter.validate()
+    params.id = params.id || hash.randomString(12).toLowerCase()
+    params.created_by_id = userId
+    await usersFusedService.ensureUserSyncedFromToken(req)
+    var project
+    if (params.project_id) {
+      project = await projectsService.getById(params.project_id)
+    }
+    if (params.project_external_id) {
+      project = await projectsService.getByExternalId(params.project_external_id)
+      // TODO: request project from Arbimon if not found
+      params.project_id = project.id
+    }
+    if (project) {
+      const allowed = await hasPermission(UPDATE, req.rfcx.auth_token_info, project.id, PROJECT)
+      if (!allowed) {
+        throw new ForbiddenError('You do not have permission to create stream in this project.')
       }
-      if (convertedParams.project_external_id) {
-        const externalProject = await projectsService.getByExternalId(convertedParams.project_external_id)
-        const allowed = await hasPermission(UPDATE, req.rfcx.auth_token_info, externalProject.id, PROJECT)
-        if (!allowed) {
-          throw new ForbiddenError('You do not have permission to create stream in this project.')
-        }
-        convertedParams.project_id = externalProject.id
-      }
-      convertedParams.id = convertedParams.id || hash.randomString(12)
-      convertedParams.created_by_id = userId
-      return streamsService.create(convertedParams, { joinRelations: true })
-    })
-    .then(streamsService.formatStream)
-    .then(stream => res.location(`/streams/${stream.id}`).status(201).json(stream))
-    .catch(httpErrorHandler(req, res, 'Failed creating stream'))
+    }
+
+    stream = await streamsService.create(params, { joinRelations: true })
+    if (arbimonService.isEnabled && source !== 'arbimon') {
+      const idToken = req.headers.authorization
+      var arbimonSite = await arbimonService.createSite(stream.toJSON(), idToken)
+      stream = await streamsService.update(stream, { external_id: arbimonSite.site_id }, { joinRelations: true })
+    }
+    res.location(`/streams/${stream.id}`).status(201).json(streamsService.formatStream(stream))
+  } catch (e) {
+    httpErrorHandler(req, res, 'Failed creating stream')(e)
+    if (stream) {
+      streamsService.del(stream, { force: true })
+    }
+  }
 })
 
 /**
@@ -118,6 +129,10 @@ router.post('/', function (req, res) {
  *           enum:
  *             - me
  *             - collaborators
+ *       - name: updated_after
+ *         description: Only return streams that were updated since/after (iso8601 or epoch)
+ *         in: query
+ *         type: string
  *       - name: start
  *         description: Limit to a start date on or after (iso8601 or epoch)
  *         in: query
@@ -140,6 +155,12 @@ router.post('/', function (req, res) {
  *         in: query
  *         type: int
  *         default: 0
+ *       - name: sort
+ *         description: Name of field to sorted / "-" for DESC default for ASC
+ *         in: query
+ *         type: string
+ *         example: is_public,-updated_at
+ *            
  *     responses:
  *       200:
  *         description: List of streams objects
@@ -166,11 +187,13 @@ router.get('/', (req, res) => {
   params.convert('is_public').optional().toBoolean()
   params.convert('is_deleted').optional().toBoolean()
   params.convert('created_by').optional().toString().isEqualToAny(['me', 'collaborators'])
+  params.convert('updated_after').optional().toMomentUtc()
   params.convert('start').optional().toMomentUtc()
   params.convert('end').optional().toMomentUtc()
   params.convert('keyword').optional().toString()
   params.convert('limit').optional().toInt().default(100)
   params.convert('offset').optional().toInt().default(0)
+  params.convert('sort').optional().toString()
 
   return params.validate()
     .then(async () => {
@@ -262,30 +285,38 @@ router.get('/:id', (req, res) => {
  *       404:
  *         description: Stream not found
  */
-router.patch('/:id', hasStreamPermission('U'), (req, res) => {
-  const streamId = req.params.id
-  const convertedParams = {}
-  const params = new Converter(req.body, convertedParams)
-  params.convert('name').optional().toString()
-  params.convert('description').optional().toString()
-  params.convert('is_public').optional().toBoolean()
-  params.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
-  params.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
-  params.convert('altitude').optional().toFloat()
-  params.convert('restore').optional().toBoolean()
+router.patch('/:id', hasStreamPermission('U'), async (req, res) => {
+  try {
+    const streamId = req.params.id
+    const source = req.headers.source
+    const converter = new Converter(req.body, {})
+    converter.convert('name').optional().toString()
+    converter.convert('description').optional().toString()
+    converter.convert('is_public').optional().toBoolean()
+    converter.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
+    converter.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
+    converter.convert('altitude').optional().toFloat()
+    converter.convert('restore').optional().toBoolean()
 
-  return params.validate()
-    .then(() => usersFusedService.ensureUserSyncedFromToken(req))
-    .then(() => streamsService.getById(streamId, { includeDeleted: convertedParams.restore === true }))
-    .then(async stream => {
-      if (convertedParams.restore === true) {
-        await streamsService.restore(stream)
+    const params = await converter.validate()
+    await usersFusedService.ensureUserSyncedFromToken(req)
+    const stream = await streamsService.getById(streamId, { includeDeleted: params.restore === true })
+    if (params.restore === true) {
+      await streamsService.restore(stream)
+    }
+    const updatedStream = await streamsService.update(stream, params, { joinRelations: true })
+    if (arbimonService.isEnabled && source !== 'arbimon') {
+      try {
+        const idToken = req.headers.authorization
+        await arbimonService.updateSite(updatedStream.toJSON(), idToken)
+      } catch (err) {
+        console.error('Failed updating stream in Arbimon', err)
       }
-      return streamsService.update(stream, convertedParams, { joinRelations: true })
-    })
-    .then(streamsService.formatStream)
-    .then(json => res.json(json))
-    .catch(httpErrorHandler(req, res, 'Failed updating stream'))
+    }
+    res.json(streamsService.formatStream(updatedStream))
+  } catch (e) {
+    httpErrorHandler(req, res, 'Failed updating stream')(e)
+  }
 })
 
 /**
@@ -310,11 +341,24 @@ router.patch('/:id', hasStreamPermission('U'), (req, res) => {
  *       404:
  *         description: Stream not found
  */
-router.delete('/:id', hasStreamPermission('D'), (req, res) => {
-  return streamsService.getById(req.params.id, { joinRelations: true })
-    .then(streamsService.softDelete)
-    .then(json => res.sendStatus(204))
-    .catch(httpErrorHandler(req, res, 'Failed deleting stream'))
+router.delete('/:id', hasStreamPermission('D'), async (req, res) => {
+  try {
+    const source = req.headers.source
+    const stream = await streamsService.getById(req.params.id, { joinRelations: true })
+    const streamId = stream.id
+    await streamsService.del(stream)
+    if (arbimonService.isEnabled && source !== 'arbimon') {
+      try {
+        const idToken = req.headers.authorization
+        await arbimonService.deleteSite(streamId, idToken)
+      } catch (err) {
+        console.error('Failed deleting site in Arbimon', err)
+      }
+    }
+    res.sendStatus(204)
+  } catch (e) {
+    httpErrorHandler(req, res, 'Failed deleting stream')(e)
+  }
 })
 
 module.exports = router